[CPU]Remove warning suppression '-Wno-sign-compare'  and fix warnings (#16476)
* fix some warning sign-compare
1: get_length() sometime return dimension, sometime return static_dimension(template param), all convert to int64_t for comparing;
	src/core/include/openvino/core/dimension.hpp:28 value_type=int64_t;
	src/plugins/intel_cpu/src/utils/shape_inference/static_dimension.hpp:24 value_type = size_t;
2: auto default = int; for loop and const declaration;
3: sign and no-sign compare, compiler will convert them to no-sign to compare automatically;
	So recommend to static_cast<size_t>;


* -    size_t brg0BaseIdx = -1;
+    int64_t brg0BaseIdx = -1;

And some test case sign no-sign comparison.

* Fix new warnings after rebase.


* Update based on maxnick review.


* Process latest comments.


* Change embIndex from int to size_t


* 1: Revert lastNumSegments_, segmentIds_ to int
2: Initialize size_t brg1BaseIdx = std::numeric_limits<size_t>::max();


* Simplify this comparing.
-            indexes.push_back(std::max<size_t>(idx - 1, 0u));
+            indexes.push_back(0u == idx ? 0 : idx - 1);


* Fix arm build warning.


* Fix ARM plugin build fail issue.


* Fix new warning


* outConf.inPlace() maybe is -1


---------

+++ src/plugins/intel_cpu/src/nodes/adaptive_pooling.cpp
@@ -140,8 +140,8 @@ void AdaptivePooling::getSupportedDescriptors() {
 
 bool AdaptivePooling::needShapeInfer() const {
     const auto newSpatialDimsPtr = reinterpret_cast<int32_t *>(getParentEdgesAtPort(1)[0]->getMemoryPtr()->GetPtr());
-    for (size_t i = 0; i < spatialDimsCount; i++) {
-        if (spatialDimsValue[i] != newSpatialDimsPtr[i]) {
+    for (int i = 0; i < spatialDimsCount; i++) {
+        if (static_cast<int32_t>(spatialDimsValue[i]) != newSpatialDimsPtr[i]) {
             for (size_t j = 0; j < spatialDimsValue.size(); j++) {
                 spatialDimsValue[j] = newSpatialDimsPtr[j];
             }
@@ -211,7 +211,7 @@ void AdaptivePooling::execute(dnnl::stream strm) {
     const auto *srcPooledSpatialShapes = reinterpret_cast<const int *>(getParentEdgeAt(1)->getMemoryPtr()->GetPtr());
     auto *dst = reinterpret_cast<float *>(getChildEdgeAt(0)->getMemoryPtr()->GetPtr());
 
-    if (srcMemory1.GetShape().getElementsCount() != spatialDimsCount)
+    if (static_cast<int>(srcMemory1.GetShape().getElementsCount()) != spatialDimsCount)
         IE_THROW() << errorPrefix << "has input spatial dimension (" << srcMemory1.GetShape().getElementsCount()
                    << ") inconsistent with pooling vector size (" << spatialDimsCount << ")";
